UefiPayloadPkg/ElfLoader: Support loading DYN ELF image
DYN ELF image is built when "-pie" is supplied to linker.
DYN ELF image contains the dynamic section that points to
a specific relocation section. This relocation section is supposed
to be used by loader to do runtime relocation.

Enhance the ELF loader to support using this relocation section
for runtime relocation.

Signed-off-by: Ray Ni <ray.ni@intel.com>
Cc: Maurice Ma <maurice.ma@intel.com>
Cc: Guo Dong <guo.dong@intel.com>
Cc: Benjamin You <benjamin.you@intel.com>