UefiPayloadPkg: Let Universal Payload support CBOR as intreface
To follow USF spec 1.0, add support for Universal Payload to accept CBOR
as intreface. The macro UNIVERSAL_SCALABLE_FIRMWARE_REVISION_10 is added.
When building Universal Payload, if UNIVERSAL_SCALABLE_FIRMWARE_REVISION_10
is set as TRUE, a CBOR interface will be accepted, else, Hoblist is used.

PayloadLoaderPeim can check the SpecRevision in upld_info
if the SpecRevision is equal or larger than 1.0, section will be reported
in cbor dataformat, else Guid Hob will be created to contain the section
information

Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>